+++ distributed-systems/src/raft/log/raft_log_entry.go
@@ -2,6 +2,6 @@ package log
 
 type RaftLogEntry struct {
 	value uint32
-	term  uint32
+	term  uint64
 	index uint32
 }
